Move starship to its own component
