+++ components/index.jsx
@@ -1,2 +1,3 @@
 export { Cart } from './Cart';
 export { AddToCart } from './AddToCart';
+export { Starship } from './Starship';
